optimizations and try to fix ci
