+++ Cargo.toml
@@ -7,5 +7,22 @@ license = "MIT"
 publish = false
 repository = "https://github.com/Whist-Team/Whist-Browser"
 
+[profile.dev.package."*"]
+opt-level = 3
+
+[profile.dev]
+opt-level = 1
+
+[profile.release]
+opt-level = "s"
+lto = "thin"
+codegen-units = 1
+
+[features]
+dev = [
+    "bevy/dynamic"
+]
+
 [dependencies]
-bevy = "0.7.0"
+console_error_panic_hook = "0.1"
+bevy = { version = "0.7.0", features = ["wayland"] }
